pyproject.toml
[tool.poetry]
name = "service_api"
version = "0.1.0"
description = "Simple service orders management API"
authors = ["Isaque Alves <isaquealves@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.1.2"
flask-restful = "^0.3.8"
flask-orator = "^0.2.0"
psycopg2 = "^2.8.5"
python-decouple = "^3.3"
connexion = "^2.6.0"
cryptography = "^2.9.2"
marshmallow = "^3.5.1"
flask-marshmallow = "^0.12.0"


[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
pudb = "^2019.2"
pytest-pudb = "^0.7.0"
coverage = "^5.1"
flask-script = "^2.0.6"
docopt = "^0.6.2"
bandit = "^1.6.2"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
